rename: day util 경로 변경
